[FEAT/#5] 로그아웃 버튼 클릭 시 다이얼로그 창으로 다시 물어보기

+++ app/src/main/java/org/android/go/sopt/ui/main/mypage/MyPageFragment.kt
@@ -3,12 +3,12 @@ package org.android.go.sopt.ui.main.mypage
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
+import androidx.appcompat.app.AlertDialog
 import org.android.go.sopt.GoSoptApplication
 import org.android.go.sopt.R
 import org.android.go.sopt.binding.BindingFragment
 import org.android.go.sopt.databinding.FragmentMyPageBinding
 import org.android.go.sopt.ui.login.LoginActivity
-import org.android.go.sopt.ui.main.MainActivity
 
 class MyPageFragment : BindingFragment<FragmentMyPageBinding>(R.layout.fragment_my_page) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -18,25 +18,34 @@ class MyPageFragment : BindingFragment<FragmentMyPageBinding>(R.layout.fragment_
     }
 
     private fun initUserProfile() {
-        val name = GoSoptApplication.prefs.getString(NAME_KEY, null)
-        val hobby = GoSoptApplication.prefs.getString(HOBBY_KEY, null)
+        val name = GoSoptApplication.prefs.getString(R.string.prefs_name_key, null)
+        val hobby = GoSoptApplication.prefs.getString(R.string.prefs_hobby_key, null)
         binding.tvName.append(name)
         binding.tvHobby.append(hobby)
     }
 
     private fun initLogoutButtonClickListener() {
         binding.btnLogout.setOnClickListener {
-            GoSoptApplication.prefs.deleteAllData()
-
-            Intent(requireContext(), LoginActivity::class.java).apply {
-                addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK)
-                startActivity(this)
-            }
+            showAlertDialog()
         }
     }
 
-    companion object {
-        private const val NAME_KEY = "name"
-        private const val HOBBY_KEY = "hobby"
+    private fun showAlertDialog() {
+        val builder = AlertDialog.Builder(requireContext())
+        builder.setMessage(R.string.dialog_logout_msg)
+            .setPositiveButton(R.string.dialog_yes) { dialog, id ->
+                GoSoptApplication.prefs.deleteAllData()
+                navigateToLoginScreen()
+            }
+            .setNegativeButton(R.string.dialog_no, null)
+            .create()
+        builder.show()
+    }
+
+    private fun navigateToLoginScreen() {
+        Intent(requireContext(), LoginActivity::class.java).apply {
+            addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK)
+            startActivity(this)
+        }
     }
 }
